Updated openstack/openstack Project: openstack/nova  2c3256c4a0492c05176a2409f0cb1e1d440c253e
Split out part of map_cell_and_hosts to return a uuid

Top level nova-manage methods need to return an exit code. We need a way
to run the logic of map_cell_and_hosts and get access to the cell
mapping uuid that's used. This splits that logic to a private method
that can be shared.

The case where all hosts are already mapped to the cell has been
modified to return the cell mapping uuid rather than 0. In that case
there may still be unmapped instances so getting access to that uuid is
beneficial.

Change-Id: Ia2d82243fd8e92f270801bcad2222808f2ee1a98