Updated openstack/openstack Project: openstack/nova  7dc5d7222c4b9f0ebfd185ee1ced6a10a8ec8d18
Rename Raw backend to Flat

As mentioned in a comment (which this patch deletes), calling this
class 'Raw' was confusing, because it is not always raw. It is also a
source of bugs, because some code assumes that the format is always
raw, which it is not. This patch does not fix those bugs.

We rename it to 'Flat', which describes it accurately. We also add
doctext describing what it does.

DocImpact

The config option libvirt.images_type gets an additional value:
'flat'. The effect of this is identical to setting the value 'raw'.

Change-Id: I93f0a2cc568b60c2b3f7509449167f03c3f30fb5