Updated openstack/openstack Project: openstack/keystone  6d07233f53f5497c3aeecf437db697097f5fabc7
Move the revoke abstract base class out of core

This patch moves the revoke abstract base class out of core and into
backends/base.py

This removes dependencies where backend code references code in the
core. The reasoning being that the core should know about the backend
interface, but the backends should not know anything about the core
(separation of concerns). And part of the risk here is a potential for
circular dependencies.

Partial-Bug: #1563101

Change-Id: I0550e0fedfb28b16e9803d225df4d2337e6c65e6