Updated openstack/openstack Project: openstack/tripleo-puppet-elements  3b52cd4b4bfeb9b3f513db31083cf4ed01a8b09a
Honor $DIB_DEFAULT_INSTALLTYPE in puppet-modules

Makes the puppet-modules element honor the $DIB_DEFAULT_INSTALLTYPE
environment variable. This environment variable is set to source by
default.

Other elements, such as the enable-packages-install element may set it
to package.

Since the enable-packages-install element sets it to package at
01-export-install-types.bash[1], 01-puppet-modules-install-types.sh is
moved to 02-puppet-modules-install-types.sh so that it runs later.

Closes-Bug: 1583263
[1] https://github.com/openstack/tripleo-image-elements/blob/master/elements/enable-packages-install/environment.d/01-export-install-types.bash

Change-Id: I1d14fc9b113d45e7ab613f44011f442e4436a904