Updated openstack/openstack Project: openstack/python-troveclient  df479520d42cdb40cc52e16ac67f5a55532a617a
Make username and password non-required params

If a user is passing in a keystoneauth Session object to the session
parameter, Session is the thing that should be handling auth
interactions, and as such troveclient does not need to know the username
and password. In fact, for a given session they may not be values that
exist, depending on keystone auth plugin. Further, they are merely
passthrough parameters to _construct_http_client which does not require
them.

Change-Id: I3b71c79a90736eef5d0951bb6e7b8a4d95ec015f