Updated openstack/openstack Project: openstack/nova  134b52e1a0c8bca671789b9b7e0b9cc4e1c807d3
ironic: fix call to _cleanup_deploy on config drive failure

The call to _cleanup_deploy when config drive generation failed
during spawn didn't match the method signature. This was missed
in unit testing because the assertion on the mock of that method
matched the actual call, but not the actual method signature.

This fixes the call and also fixes the test by auto-spec'ing the
_cleanup_deploy method in the mock so that it validates the actual
function signature is called correctly.

In order to use autospec properly here, the mock has to be on the
driver object rather than the class.

Change-Id: Ic2c096ef846f11f94aa828222c927ed7d03051c9
Closes-Bug: #1581246