Updated openstack/openstack Project: openstack/horizon  f31634752caa3350a64b85465410b216243e3043
Add recursive deletion to Swift UI

This patch adds client-side recursive delete (there is
no server-side facility for this) so that users may
delete non-empty folders. This also allows for select-all
and row-select to select folders, and delete-selected
to work on those folders, regardless of whether they are
empty or not.

Change-Id: Ibad1342945d330ffc7d8f8af68156143716d435c
Partially-Implements: blueprint swift-ui-functionality
Co-Authored-By: Timur Sufiev <tsufiev@mirantis.com>