Updated openstack/openstack Project: openstack/debtcollector  cefe6179068f545d755bebd26937d3745a694880
Add tests for decorated argspec preservation

Tests are added that show that several of the decorators, when
used on functions, lose the argspec. This causes the documentation
to be generated incorrectly since the args aren't shown for
decorated functions.

The tests show that some of the decorators work as expected.

Change-Id: Id024f5110082a88c30e71e991764320b3aed07aa