Updated openstack/openstack Project: openstack/heat  dff591775d361c757eca3c4f619be39f18d8b5a7
Use original request context for logging

The various Oslo libraries call oslo_context.context.get_current() to get
the current context for various purposes such as logging. This is a
thread-local variable, so this patch ensures that it is set in the
asynchronous threads that we spawn for various long-running tasks. It also
ensures that we pass overwrite=False when creating new contexts (e.g.
reconstructing a stored context from a stack in the database) so that we
don't overwrite the thread-local variable.

Change-Id: Ic81eda7cb1dddfa72421119a6fe26a093eee3aab