Updated openstack/openstack Project: openstack/horizon  1e4f5d1dad685e437de9864064e33ca2bb7ddb28
Modify default sort in flavor table in ng-instance

Default sort on the flavor table is specified as RAM (only the case
of available table). But at the moment, it doesn't work.
It is caused by undefined variables (isAvailableTable).
This variables seems to be a flag wheather a table is Allocated table
or Available table. However, these tables are implemented separately.
Thus, it doesn't need to use this variable. Therefore, this patch
will remove the condition using this variable.

Change-Id: I912850d352e547082bb25c5fced938d8cdc50dc9
Closes-Bug: #1580056