Updated openstack/openstack Project: openstack/ironic  9739a22857a95fb7e35831b0a75c608b801c8bb5
Remove link to Liberty configs

This removes the link to the Liberty configs from our documentation.
Our documentation should reflect what is on the master branch, but
since the configuration documentation is only available on a (named)
release basis, we should only link to the latest, which is Mitaka
configs.

Change-Id: I2d4c857dcbe3745ef8aad714600ea32c9762a628