Updated openstack/openstack Project: openstack/tripleo-heat-templates  05720272733a8fa62713411673b233b416f3f3a1
Add net-config-static.yaml

Add a new network config for configuring a static IP on a single
interface.  Similar to net-config-static-bridge.yaml, except that it
does not create a bridge and add the specificed interface to that
bridge.

Where net-config-static-bridge.yaml would be used as the controller
network config for the scenario of configuring a static IP,
net-config-static.yaml would be used for the compute node when you want
a static IP, but do not need a bridge.

Change-Id: Ide3be3b90fcf5e4b6045be3128dfb5c812fbb2e8