Updated openstack/openstack Project: openstack/cinder  5e07b872c0a43a34e9601d4ca8151622ef84622a
RBD: include cluster name in connection info

This additional field will be ignored for now, but it will eventually be
used in os_brick by the RBD connector.

The os_brick project does not have configuration options, so this
approach is required to get cluster configuration settings into the
connector.

Change-Id: I6538c81a4d90d3677b46682ece30a97bc4f11423