Updated openstack/openstack Project: openstack/nova  54d66ee894cb487aa2ff3c394862cb9c3763443d
Make neutronapi get_floating*() methods return objects

This makes the neutronapi module return NeutronFloatingIP objects from
the relevant methods, instead of bare dicts.

Note that the API is returning string floating ip identifiers when neutron
is in use, as opposed to the integer ones returned when nova-network is in
use. This patch does not address that problem, but uses the new
NeutronFloatingIP.

Needed for blueprint rm-object-dict-compat-newton

Co-Authored-By: Ryan Rossiter <rlrossit@us.ibm.com>
Change-Id: Ie60fb8661195eec9c01a57129efe8145504789e1