Updated openstack/openstack Project: openstack/cinder  6375213abe564aa5ce2cb09b33e726b8015aef77
Prevent Multitable conditional updates

Since PostgreSQL doesn't support multitable updates we want to raise an
exception whenever we try to do such queries, even if they would have
succeeded with MySQL.

That way, if we have a multitable query in our code base that is not
getting tested in our PostgreSQL CI job, at least it will always fail on
devstack deployments even when they use the MySQL DB engine when the
query gets executed.

Without this patch we just wouldn't see the issue on our devstack
deployment that uses MySQL, as multitable operations are supported.

Change-Id: Ib45d3fb6b6c580615f4685e3d6f70705f1afffe9