Updated openstack/openstack Project: openstack/nova  91a534b0441a8529dcd67b5e185959369e51f963
Add tests for floating_ip private functions

There are 2 try/except blocks within _translate_floating_ip_view() that
cause in different resulting dictionaries. These situations are
untested, so now we can test for different scenarios of missing
attributes on the floating IP. An if check was also added to the
function because the types passed in are objects, so if the attribute is
not set on the object, it raises a NotImplementedError. So instead of
catching that too, we check to see if the key is in the object (which
checks if it is set).

Change-Id: Ia5739f214f57525b3083ef3f201f6f08198703e5