Updated openstack/openstack Project: openstack/nova-specs  bd804da21850dfee0fa79576d72ebd7ace3ab1a1
Complete Nova support for Glance's v2

Glance v1 is subject to deprecation, we need Nova to move away
from v1 and use v2 fully. This spec discusses how we plan to
make that transition. This also documents the testing plan
for both v1 and v2 and few limitations associated with it.

Previously-approved: Mitaka

Change-Id: I1247b436f9cbb21eb6e3d5cbabe5345e6037ceb2
Co-authored-by: Flavio Percoco <flaper87@gmail.com>
Co-authored-by: Sudipta Biswas <sbiswas7@in.ibm.com>