Updated openstack/openstack Project: openstack/nova  813787644bd11ffb8bdf46a547bd25982d995dea
api-ref: complete verification of baremetal api

This cleans up the baremetal proxy, which was horribly inaccurate (the
paths weren't right after the method verification)

A warning is stuck in the preamble, we really don't want people using
this. Parameters are documented as best as we could from Nova
source. Honestly, people should just talk to ironic directly and use
their docs.

Part of bp:api-ref-in-rst

Change-Id: I66e85f414e5f0edbf052ba56691294a94615c619