Updated openstack/openstack Project: openstack/nova-specs  8d16fc11ee6d01b5a9fe1b8b7ab7fa6dff460e2a
Add Keystone project validation to quota and flavor management

Currently, Nova does not check the validity of what the user provides for the
project id value in quota and flavor management activities. This leads to
inaccurate quota entries in the database, causing the user to think Nova isn't
reporting quota information accurately. The solution proposed in this blueprint
is to add project id validation specifically to quota and flavor management
through Keystone, to provide the user with an error if their specified project
id is invalid, and to ensure quota and flavor management data is not incorrectly
modified.

bp validate-project-with-keystone
Previously-approved: kilo

Change-Id: I631cef639e9485a36ab0e605aef0598f786332dc