Updated openstack/openstack Project: openstack/oslo.policy  65f93657dc55701571a6c17aa5b7a7f63489af7f
Add policy registration and authorize method

A new RuleDefault class has been added which can be used to register
policies that will be used by a consumer of oslo.policy. These policies
are merged with those defined in policy files, with the file definitions
overriding the defaults

Registering a policy with the same name twice is considered an error and
will raise a DuplicateRuleDefaultError exception.

To facilitate projects wishing to ensure that all policies are
registered before use an authorize method has been added which errors
when checking an unregistered policy. If the policy has been registered
then the enforce method is called in the normal manner.

Change-Id: I3b6423aeed8ae80e8bf73dbda0f63ef379ccef43
Implements: bp policy-in-code