Updated openstack/openstack Project: openstack/nova  7b35aa97d84acaa503083a85c96bdf90a4aabe05
Add online migration to move keypairs from main to API database

This moves keypairs from the main database to the API database in
an online migration. Note that it refuses to run until all of the
instances have keypair information, so that we don't strand instances
that might need to lazy-load those things.

Related to blueprint cells-keypairs-api-db

Change-Id: I97b72ae3e7e8ea3d6b596870d8da3aaa689fd6b5