Updated openstack/openstack Project: openstack/cinder  94f1fe1e4663d0420476140070a499ab40b9c8b4
Add SnapshotStatus enum field

This change adds a new enum and field, SnapshotStatus and
SnapshotStatusField, that will hold the constants for the 'status'
field of the SnapshotStatus object. This enum and field are based
off the base oslo.versionedobjects enum and field. This also changes
over the Snapshot object to use the new field and bumps up the version
so newer versions know to enforce valid values. Finally, all uses of
strings for comparison and assignment to this field are changed over
to use the constants defined within the enum.

Change-Id: I968ad7a1f422811eaf437af435361dac7915b594
Partial-Implements: bp cinder-object-fields