Updated openstack/openstack Project: openstack-infra/subunit2sql  aaa17dcbf3a31f38e6eb0cd93023544ff316f4d1
Add link to published release notes in readme

One of the things missing from the README was a link to the release
notes. We used to have this when the changelog was manually curated
in the readme, but since we switched to an autogenerated short log
and reno the readme page doesn't contain any release info itself.
This commit adds the links to point people to where these docs live.

Change-Id: Ie8494a9eec5ee2a2ff0ffdb417fbee119e77f047