Updated openstack/openstack Project: openstack/neutron  73d7018f9e98f5258a93833dd8ecc831c430c4cf
Deprecate neutron-debug

We will not follow the typical deprecation cycle and thus we will
not remove this in the release after the deprecation is advertised.

Nonetheless we want to signal the need for a more comprehensive set
of tools and APIs that will eventually make people want to move away
from the limitations of neutron-debug.

Closes-bug: #1583700

Change-Id: Ibe471f186b9b5c9c15c5857a35cbaca6b8d284d6