Updated openstack/openstack Project: openstack-infra/subunit2sql  58b2b6b526a595bfaf06072a4f7dc25569542881
Remove cap on oslo.db version

This commit removes the requirements cap on oslo.db. This was
originally added because of the removal of a private interface that
subunit2sql depended on in it's alembic env.py. A public interface
was eventually added to provide analogous functionality but
subunit2sql never removed the cap and used the new interface;
until now.

Change-Id: Idbc17096d0d3fe43261026b866041ae8a97d86fc