Updated openstack/openstack Project: openstack/neutron  9c65b8ee80091a75e89303476d988b20201b16fa
Add method to get service provider names by resources

This adds a method to the service type framework to retrieve
the names of the resource providers associated with a list
of resource IDs. This will be used by the L3 flavor framework
to lookup all of the associated drivers for a list of routers.

Partially-Implements: bp/multi-l3-backends
Change-Id: I4f0d7df493c5a5a7b00dc6f43d3827536a474a40