Updated openstack/openstack Project: openstack/nova  1528db4cbe4e9cc6f20acff29b64e6ba39f99c12
Only attempt to inject files if the injection disk exists

_inject_data was checking for the existence of the image to inject
into. However, the only caller of _inject_data has just created that
image. We move the _inject_data call up to where the image is created,
and pass the existing image object in directly. We modify _inject_data
to remove the existence check as, bar a delete race which was present
anyway, we know it exists.

Change-Id: Icb926e64151048c6ce8353de604417d5d79858e4