Updated openstack/openstack Project: openstack/ironic  5123c755a90a3a5ad7c3f18ce5ed69b3e176777a
Read the Sphinx html_last_updated_fmt option correctly in py3

The html_last_updated_fmt option is the output of a shell command, which
is interpeted as a byte string in python 3 and subsequently breaks
Sphinx.

This change converts it to a utf-8 string.

Incidentally, this also removes the quotation marks to make the
formatted output cleaner.

Change-Id: I9aa6446a24d2f340246458dea0c34d456b3eade3
Closes-bug: #1583721