Updated openstack/openstack Project: openstack/neutron  293e6bfd0fa28fb37c51c46a2392ce4996c84561
DVR: Moving router from dvr_snat node removes the qrouters

Removing the router from dvr_snat node removes the qrouters
that are servicing the VM and dhcp ports.

If there are still dvr serviceable ports in the dvr_snat node,
and if router remove command is executed to remove the router
association from the dvr_snat agent then only the snat
functionality should be moved to the the different agent
and the router namespaces should be untouched.

This patch checks if there are any dvr serviceable ports for
the dvr router and if exists, it will not send a router_remove
message to the agent, but instead will send an router_update
message to the agent.

Change-Id: I5a3ba329346ab0d5ea7b0296ec64cc8e5fb4056d
Closes-Bug: #1564575