Updated openstack/openstack Project: openstack/django_openstack_auth  086fc270fa41994fefe1212354c194b91c915179
Use login endpoint as key for AVAILABLE_REGIONS

The endpoint defined in request.user.endpoint may differ from the
endpoint selected from Horizon's AVAILABLE_REGIONS. If so, this will
result in the region appearing as 'None' in Horizon.

This will use the login endpoint as the key for setting the region.

Change-Id: I02d8069d2c8dcb5c24950279b1e40469072bf3bd
Closes-Bug: 1494287