Updated openstack/openstack Project: openstack/nova  f6d341ca5849322d4832cb7a08c07dd90fe83078
HyperV: Add SerialConsoleOps class

This patch introduces a class that will be the entry point for
VM serial console related operations.

This is responsible of setting up serial console handlers for each
instance. This will be used for retrieving the VM console logs as
well as interactive serial console access.

Partially-implements: blueprint hyperv-serial-ports

Change-Id: I40a39e808a63063d37eca8c7fec484d0089987de