Updated openstack/openstack Project: openstack/heat  d253507a8239ec58ab54b5a226c23ec3e54a5b44
Don't use stored context to reset stacks

The resetting of in-progress stacks at startup is done using an admin
context. No user actions (e.g. ReST API calls) are performed, so there is
no need to load the user context that is stored with the stack. (Nor is the
current context rewritten to the database in the process of resetting the
state.)

Since loading the user context from the DB may result in a call to the
keystone API, eliminating this ensures that stacks can be reset
successfully even if heat-engine starts before keystone is available.

Change-Id: I8109c622b7661254a658d78d98f8dc8f756d8755
Closes-Bug: #1570569