Updated openstack/openstack Project: openstack/horizon  5001e89b34aa7f37187f43276563313c2b2a0c79
Fix the issue that delete actions are unclickable

In detail page like volume, image and so on, delete action
is not clickable. Delete actions are set by the code below.
https://github.com/openstack/horizon/blob/master/horizon/static/horizon/js/horizon.forms.js#L366
However, because of the change below, these no longer match this selector.
https://review.openstack.org/#/c/284421/18
This patch will fix it so as to match the selector above.

Change-Id: Ia4ef05b8811c09dedeeb4c3349e989c4e3af5122
Closes-Bug: #1580975