Updated openstack/openstack Project: openstack/requirements  2e015bfe0e1e3e969ac041e67ecfec6669388fb8
Drop python-ldap from global requirements

With keystone moving to pyldap (and ldappool requiring pyldap) there is
no longer a dependency on python-ldap in OpenStack as far as I have
been able to determine. This patch drops python-ldap in favor of
pyldap which is python3 compatible and a drop-in replacement for
python-ldap (it was forked from pyton-ldap).

Change-Id: Iff153146c2e94842d203204359477511028ecdbc