Updated openstack/openstack Project: openstack/requirements  b7deaa488736a04897c582706431482121ebb3fc
Bump aodhclient to 0.5.0

This change is because the aodhclient-0.1.0 is the first initial version
of aodhclient, it only provided some primitive functionalities. Until
aodhclient-0.5.0, it has provided a set of relatively well-integrated
functionalites. We are going to port aodhclient to Heat with change[1], so
it is better to bump aodhclient to version 0.5.0.

[1] I35803c0d5f43b214dda7abdcd69a33b4cd0c2646

Change-Id: I083c9ddbfa4f0f98da20900cb8eeaca38eb4d07b