Updated openstack/openstack Project: openstack/trove-specs  c1904f7241a8da3aedd24cc063aab478184404ab
Fix tests to work with current release

This change dynamically picks the release directory
based on the fact that they are in alphabetical order.
This should avoid having to modify the current_release
variable at the beginning of each cycle.

Fixes errors in trove-image-build.rst since the
tests weren't being run in the right directory.

Change-Id: If8f84d25fbdde14103db56304eba92abc93dc4ee