Updated openstack/openstack Project: openstack/nova  e05944d783368eb2c092f7894ceb8c9fc996f05f
Remove useless test of incorrect stats value

There was a test class in test_resource_tracker.py that was testing
whether if a virt driver returned a non-dict value of the 'stats' key of
the returned dict from the virt driver get_available_resource() call
that calling update_available_resource() would raise a ValueError. Well,
duh, of course it would. But why do we need to *assert* that this
behaviour occurs? It's just bloated unit tests for no reason.

Change-Id: Ie18dd38d0398af96ff805dcd3fcd18f8bc3c65f8