Updated openstack/openstack Project: openstack/nova  96e7773d1a36f15cf97f442f58f4d49aa55cf493
Remove StatsDicTestCase from test_resource_tracker

The StatsDictTestCase in test_resource_tracker.py was supposed to be
verifying that the "stats" dict field on the ComputeNode object was
being populated with some keys like num_instances,
num_proj_{project_uuid), etc. Unfortunately, all this test was really
asserting was that the behaviour of the ComputeNode object when
saving/setting the stats dict converted the values from integers to
strings.

The *actual* setting of real values of stats information is already
being fully tested with unit tests that are readable and cover all
variations of the stats keys in test_tracker.TestInstanceClaim's
methods, so we remove the StatsDictTestCase from
test_resource_tracker.py.

Change-Id: Ief658c0da2503b8e9905f5da2dd4c667a3fce550