Updated openstack/openstack Project: openstack/nova  0510f1dbb574c2a209da73b045ce812995fb2f6d
Remove conductor from test_resource_tracker

The base test case class in test_resource_tracker.py was starting a
conductor service... for no reason at all. The resource tracker no
longer directly calls any conductor calls at all. Everything is done via
the objects interfaces now or via the scheduler client. So, we remove
the call to start a conductor service and the stubout of a flavor_get
call that was not in use any more.

Change-Id: I356604a9240ed53a6f0463fc1d99989ed263b826