Updated openstack/openstack Project: openstack/cinder  7341c711206b6a7dc29906b808e9f31ce5eee1b6
Fail on disabled_reasons with more than 255 chars

Current code removes leading and trailing spaces when checking disabled
reason for a service, but then stores it as non trimmed, so we could be
trying to store in the DB a string longer than 255 characters.

To preserve existing storing behavior this patch will store leading and
trailing spaces as we are doing now, but will take them into
consideration when counting maximum length.

It will also raise an error if the reason is only comprised of spaces,
just like we are doing now.

Change-Id: Iee97d8786644ad02a83c902c7780055e2c075198
Closes-Bug: #1583652