Updated openstack/openstack Project: openstack/nova  3e7e8cc770306d5a173b4213f384327f1a3114b0
Remove service crud from test_resource_tracker

The Service object is no longer coupled with the ComputeNode object and
is not used in the resource tracker at all. test_resource_tracker.py was
still creating a service entry in its setUp() and mocking out
service-related stuff, so we remove that.

Change-Id: I83bf0253c92f318cbc782aee4e6f411925b6a7af