Updated openstack/openstack Project: openstack/python-zaqarclient  67f8f5c7220cc274f3f4f2c6822278f2cfef83ac
Refactor queue create

Now we have two key parameters for queue's constructor: auto_create
and force_create. Though we're trying to make the name can explain
itself, we're stilling missing a case when the foce_create is True
but auto_create is Flase. So this case, the queue should be created
as well. This patch fixes it.

Change-Id: I4beee2c2d01cab1988e8b28e220996336431dba1