Updated openstack/openstack Project: openstack/cinder  2d7d81df5d338cbc18b1841ad50128a4fedf0e9a
Windows iSCSI: properly handle config option

The Windows iSCSI driver attempts to use the 'image_conversion_dir'
option from the configuration object passed by the manager, without
registering it again.

The issue is that by the driver is loaded, the copy of the global
CONF object will not contain this option. When the module containing
this option is imported, it gets registered under the global CONF
object, so the driver will not be able to fetch it this way.

For this reason, operations involving volume upload will fail.

This can easily be fixed by just using CONF.image_conversion_dir,
since there's no code that expects this option to be defined per
backend.

Change-Id: Icf5af04dc431e44d7ce332bf8c2fb9b0a55cdc9a
Closes-Bug: #1583993