Updated openstack/openstack Project: openstack/nova  7c8fab0ecbbc71bae294615f7e8a01c3d2955081
libvirt: Add serial ports to the migration data object

The serial console code in nova picks TCP ports for each serial
interface of an instance. When doing a live migration, the
TCP ports used for the migrating instance might be in use on
the target hypervisor which causes the live migration to fail.

This patch adds a field to the migration data object so that a follow
up patch can use that to rewrite the guest XML.

Change-Id: I970014ee0ee8059f5976797bf344fea9a7ccb32d
Related-Bug: #1455252