Updated openstack/openstack Project: openstack-infra/tripleo-ci  f92099b67369db3edcbf5d90d0d9e4f3bc90bc1d
Allow the mirror server to be publicly accessible

Developers, community members and CI jobs would like to be able to
download images that are known to have passed tripleo-ci. To make
them available we allow httpd to listen to all IP's and restrict
the upload script to only accept files from the test network.

Change-Id: I8708fd8ef069283093b7c070e86892f5abce1cb1